==> 1_Concaténation matrice temps.ipynb <==
# Concaténation des matrices temps

import pandas as pd
import warnings
warnings.filterwarnings('ignore')

df1 = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\matrice_temps_2')
df2 = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\matrice_temps_3')

rows_verif = df1.shape[0] + df2.shape[0]
df_final = pd.concat([df1, df2])
df_final.shape[0] == rows_verif

df_final.drop_duplicates(inplace=True)
df_final.shape[0]

df_final.to_csv(r'matrice_temps_complete.csv')
