==> 2_Création fichier référence décès.ipynb <==
## Création de fichiers de référence pour le nombre de décès et de crémations moyen sur une année, le nombre d'habitants, avec les niveaux de granularité suivants : région, département, commune, ainsi que la présence d'un crématoire

import warnings
warnings.filterwarnings("ignore")

### On importe le package pandas :

import pandas as pd

## Table de référence INSEE et table des décès

### On charge la table de référence géographique INSEE :

table_insee = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\table-appartenance-geo-communes-23-nettoyé.csv')

table_insee

### On charge la table INSEE répertoriant tous les décès 2022-2023 :

table_deces = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\décès_2022-2023.csv',
                          sep=';')
table_deces

### Inspectons l'état des tables :

table_insee.isnull().sum()

table_deces.isnull().sum()

### Nous avons plusieurs actions à réaliser : 
- éliminer les colonnes inutiles à notre analyse
- éliminer les valeurs manquantes dans la colonne COMDOM, qui représente la commune de domicile et est essentielle à notre analyse
- ne garder que les décès de l'année 2022 (analyse réalisée en 2023, les données 2023 sont donc incomplètes)

table_insee.drop(columns=['EPCI', 'NATURE_EPCI', 'ARR', 'CV', 'ZE2020', 'UU2020', 'TUU2017', 'TDUU2017', 'AAV2020', 'TAAV2017',
                          'TDAAV2017', 'CATEAAV2020', 'BV2022'], 
                 inplace=True)

table_insee.info()

table_deces.dropna(subset='COMDOM', 
                   inplace=True)

table_deces.isna().sum()

table_deces_2022 = table_deces[table_deces['ADEC'] == 2022]

table_deces_2022

table_deces_2022.drop(columns=['MDEC', 'JDEC', 'ANAIS', 'MNAIS', 'JNAIS', 'SEXE', 'LIEUDEC2', 'COMDEC'],
                 inplace=True)

table_deces_2022.info()

## Table répertoriant les crématoriums et projets de crématorium et table du dernier recensement

table_infrastructures = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\Infrastructures_v2_202309251942.csv')
table_infrastructures

table_recensement = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\base-pop-historiques-1876-2020.csv', header=5)
table_recensement

### Commençons par la table_recensement, qui sera la plus simple à traiter:
- Ne garder que les colonnes CODGEO, et PMUN20, REG, DEP et LIBGEO étant déjà présents dans notre table de référence géographique

table_recensement.drop(columns=['REG', 'DEP', 'LIBGEO', 'PMUN06', 'PMUN07', 'PMUN08', 'PMUN09', 'PMUN10', 'PMUN11', 'PMUN12',
                                'PMUN13', 'PMUN14', 'PMUN15', 'PMUN16', 'PMUN17', 'PMUN18', 'PMUN19', 'PSDC62', 'PSDC68',
                                'PSDC75', 'PSDC82', 'PSDC90', 'PSDC99', 'PTOT36', 'PTOT54', 'PTOT1876', 'PTOT1881',
                                'PTOT1886', 'PTOT1891', 'PTOT1896', 'PTOT1901', 'PTOT1906', 'PTOT1911', 'PTOT1921', 'PTOT1926',
                                'PTOT1931'],
                       inplace=True)
table_recensement.info()

### Pour la table_infrastructures, nous ne garderons que les colonnes code_INSEE (que nous renommerons CODGEO pour correspondre à la nomenclature INSEE), et STATUT. Nous filtrerons également sur l'année 2021 (données les plus récentes)

table_infrastructures = table_infrastructures[table_infrastructures['Annee'] == 2021]
table_infrastructures.drop(columns=['num', 'LOCALISATION', 'x', 'y', 'CREMATIONS',],
                           inplace=True)
table_infrastructures.rename(columns={'code_INSEE' : 'CODGEO'},
                             inplace=True)
table_infrastructures.info()

## Maintenant, extrayons ces tables et menons notre analyse sur PowerBI :

table_deces_2022.to_csv(r'table_deces_2022.csv')
table_insee.to_csv(r'table_insee.csv')
table_infrastructures.to_csv(r'table_infrastructures.csv')
table_recensement.to_csv(r'table_recensement.csv')
