==> 3_Filtrage_zones_influence_et_projets.ipynb <==
# Importations et nettoyage

import pandas as pd


matrice_temps = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\matrice_temps_complete.csv')
code_espelia_code_insee = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\Coordonnées_villes_par_codes_INSEE.csv')
table_insee = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\table_insee.csv')
infrastructures = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\table_infrastructures.csv')
table_recensement = pd.read_csv(r'C:\Users\Nicolas et Maëva\Desktop\Projets DataSpell\DataBird\Projet_final\table_recensement.csv')

matrice_temps['Temps_trajet'] = matrice_temps['Temps_trajet'].astype(int)
matrice_temps['Commune_destination'] = matrice_temps['Commune_destination'].astype(str)
matrice_temps['Commune_depart'] = matrice_temps['Commune_depart'].astype(str)

code_espelia_code_insee['Index'] = code_espelia_code_insee['Index'].astype(str)

table_reference = code_espelia_code_insee.merge(table_insee, left_on='INSEE_COM', right_on='CODGEO')
table_reference.drop(columns=['INSEE_COM', 'Unnamed: 0', ], inplace=True)
table_reference.rename(columns={'Index':'Index_espelia', 'CODGEO':'CODGEO_INSEE'}, inplace=True)

# Implémentation de la fonction de référencement de codes INSEE / ESPELIA

def codes(nom_de_ville) :
    row = table_reference[table_reference['LIBGEO'] == str(nom_de_ville)]
    return row

# Implémentations des différents taux de captation

def colonne_conditionnelle (ligne):
    if ligne['zone'] == 'Z1':
        return 100
    elif ligne['zone'] == 'Z2':
        return 80
    elif ligne['zone'] == 'Z3':
        return 40
    elif ligne['zone'] == 'Z4':
        return 20

# Implémentation de la fonction principale de création de zones de captation et de vérification de la présence d'un crématorium plus proche que notre projet

def zonage(code_espelia) :

    # Définition des zones de captation et création de la colonne étiquette 'zone' pour chaque zone

    Z1 = (matrice_temps[(matrice_temps['Commune_depart'] == str(code_espelia))
                        & (matrice_temps['Temps_trajet'] < 900.0)]) \
        .merge(code_espelia_code_insee, left_on='Commune_destination', right_on='Index') \
        .merge(table_insee, left_on='INSEE_COM', right_on='CODGEO') \
        .merge(table_recensement, on='CODGEO') \
        .sort_values(by='Temps_trajet') \
        .drop(columns=['Unnamed: 0_x', 'Unnamed: 0_y', 'INSEE_COM', 'Unnamed: 0', 'Commune_depart', 'Commune_destination'])
    Z1['zone'] = 'Z1'

    Z2 = (matrice_temps[(matrice_temps['Commune_depart'] == str(code_espelia))
                        & (matrice_temps['Temps_trajet'] >= 900.0)
                        & (matrice_temps['Temps_trajet'] < 1800.0)]) \
        .merge(code_espelia_code_insee, left_on='Commune_destination', right_on='Index') \
        .merge(table_insee, left_on='INSEE_COM', right_on='CODGEO') \
        .merge(table_recensement, on='CODGEO') \
        .sort_values(by='Temps_trajet') \
        .drop(columns=['Unnamed: 0_x', 'Unnamed: 0_y', 'INSEE_COM', 'Unnamed: 0', 'Commune_depart', 'Commune_destination'])
    Z2['zone'] = 'Z2'

    Z3 = (matrice_temps[(matrice_temps['Commune_depart'] == str(code_espelia))
                        & (matrice_temps['Temps_trajet'] >= 1800.0)
                        & (matrice_temps['Temps_trajet'] < 2700.0)]) \
        .merge(code_espelia_code_insee, left_on='Commune_destination', right_on='Index') \
        .merge(table_insee, left_on='INSEE_COM', right_on='CODGEO') \
        .merge(table_recensement, on='CODGEO') \
        .sort_values(by='Temps_trajet') \
        .drop(columns=['Unnamed: 0_x', 'Unnamed: 0_y', 'INSEE_COM', 'Unnamed: 0', 'Commune_depart', 'Commune_destination'])
    Z3['zone'] = 'Z3'

    Z4 = (matrice_temps[(matrice_temps['Commune_depart'] == str(code_espelia))
                        & (matrice_temps['Temps_trajet'] >= 2700.0)]) \
        .merge(code_espelia_code_insee, left_on='Commune_destination', right_on='Index') \
        .merge(table_insee, left_on='INSEE_COM', right_on='CODGEO') \
        .merge(table_recensement, on='CODGEO') \
        .sort_values(by='Temps_trajet') \
        .drop(columns=['Unnamed: 0_x', 'Unnamed: 0_y', 'INSEE_COM', 'Unnamed: 0', 'Commune_depart', 'Commune_destination'])
    Z4['zone'] = 'Z4'

    # Concaténation des zones de captation et création des colonnes 'Taux_captation' et 'Population_captée' 
    # grâce à notre fonction colonne_conditionnelle

    zonage_complet = pd.concat([Z1, Z2, Z3, Z4])
    zonage_complet['Taux_captation'] = zonage_complet.apply(colonne_conditionnelle, axis=1)
    zonage_complet['Population_captée'] = (zonage_complet['PMUN20'] * zonage_complet['Taux_captation']/100).astype(int)


    # Création de la liste des villes à vérifier

    villes_a_verif = zonage_complet.iloc[1:, :] \
        .set_index('Index')['Temps_trajet'] \
        .to_dict()


    # Pour chaque ville, on vérifie si notre projet de crématorium serait le plus proche de la ville,
    # si c'est le cas, on garde la ville, sinon on l'exclut du zonage final

    progression = 0
    villes_a_garder = []
    villes_a_garder.append(code_espelia)
    infrastructures['STATUT'].replace({'Existant':1, 'Projet':0}, inplace=True)
    for ville in villes_a_verif.keys():
        temps_a_comparer = villes_a_verif[ville]
        df = (matrice_temps[matrice_temps['Commune_depart'] == ville]) \
            .merge(code_espelia_code_insee, left_on='Commune_destination', right_on='Index') \
            .merge(infrastructures, left_on='INSEE_COM', right_on='CODGEO') \
            .sort_values(by='Temps_trajet')
        filtered_df = df.loc[df['STATUT'] == 1, 'Temps_trajet']
        if not filtered_df.empty:
            temps_crematorium_plus_proche = filtered_df.iloc[0]
        else:
            temps_crematorium_plus_proche = 99999999
        progression += 1
        print(f"Progression : {progression} / {len(villes_a_verif)}")
        if temps_a_comparer < temps_crematorium_plus_proche:
            villes_a_garder.append(ville)
    zonage_complet = zonage_complet[zonage_complet['Index'].isin(villes_a_garder)]

    return zonage_complet

# Utilisation

zonage_carcassonne = zonage((codes('Carcassonne'))['Index_espelia'].iloc[0])
zonage_carcassonne

zonage_carcassonne.to_csv(r'zonage_carcassonne')
